working condition

+++ four_matricies_to_diffusion_asymptotic.h
@@ -1,9 +1,6 @@
-#include <stdint.h>
-#include <stddef.h>
-
 void
 charge_a_b_integral(
     double* a_b_integral, double* a, double* b, double* r,
     double* A, double* B, double* K, double* I,
-    size_t n, size_t x_n, double x_delta);
+    unsigned int n, unsigned int x_n, double x_delta);
